Quara/app/src/main/java/quara/test_login/Register.java
package quara.test_login;

import android.app.AlertDialog;
import android.content.Intent;
import android.os.Bundle;
import android.support.design.widget.FloatingActionButton;
import android.support.design.widget.Snackbar;
import android.support.v7.app.AppCompatActivity;
import android.support.v7.widget.Toolbar;
import android.view.View;
import android.widget.Button;
import android.widget.EditText;

public class Register extends AppCompatActivity implements View.OnClickListener {

    Button bRegister;
    EditText etUsername, etName, etPassword;

    @Override
    protected void onCreate(Bundle savedInstanceState) {
        super.onCreate(savedInstanceState);
        setContentView(R.layout.activity_register);

        etUsername = (EditText) findViewById(R.id.etUsername);
        etName = (EditText) findViewById(R.id.etName);
        etPassword = (EditText) findViewById(R.id.etPassword);

        bRegister = (Button) findViewById(R.id.bRegister);

        bRegister.setOnClickListener(this);

    }

    private void showErrorMessage()
    {
        AlertDialog.Builder dialogBuilder = new AlertDialog.Builder(Register.this);
        dialogBuilder.setMessage("Username already exists");
        dialogBuilder.setPositiveButton("OK", null);
        dialogBuilder.show();
    }

    public void RegisterUser(User user)
    {
        ServerRequests serverRequests = new ServerRequests(this);
        serverRequests.storeUserDataInBackground(user, new GetUserCallBack() {
            @Override
            public void done(User returnUser) {
                startActivity(new Intent(Register.this, Login.class));
            }
        });
    }

    public void authenticate(User user)
    {
        ServerRequests serverRequests = new ServerRequests(this);
        serverRequests.checkUserDataInBackground(user, new GetUserCallBack() {
            @Override
            public void done(User returnUser) {
                if (returnUser == null) {
                    showErrorMessage();
                } else {
                    RegisterUser(returnUser);
                }
            }
        });
    }

    @Override
    public void onClick(View v) {
        switch(v.getId())
        {
            case R.id.bRegister:
                String name = etName.getText().toString();
                String username = etUsername.getText().toString();
                String password = etPassword.getText().toString();

                User user = new User(name, username, password);
                authenticate(user);

                break;
        }
    }
}
